+++ Observer/MessagesUpdateEvent.h
@@ -9,13 +9,13 @@
 class MessagesUpdateEvent : public Event
 {
 public:
-	MessagesUpdateEvent(int client_num, const IMessage& msg);
+	MessagesUpdateEvent(size_t client_num, const IMessage& msg);
 	virtual ~MessagesUpdateEvent() override = default;
 	virtual eventType GetEventType() const override;
-	int GetClientNum() const;
+	size_t GetClientNum() const;
 	const IMessage& GetMsg() const;
 
 private:
-	int _client_num;
+	size_t _client_num;
 	const IMessage& _msg;
 };
