+++ Server/ClientConnection/ClientConnection.h
@@ -10,8 +10,8 @@ class IMessage;
 class ClientConnection : public IObservable
 {
 public:
-	ClientConnection(std::shared_ptr<Socket> client, int number);
-	int GetNumber() const;
+	ClientConnection(std::shared_ptr<Socket> client, size_t number);
+	size_t GetNumber() const;
 	void sendStr(const std::string& str);
 	void SendMsg(const IMessage& msg);
 
@@ -22,5 +22,5 @@ public:
 private:
 	std::shared_ptr<Socket> _client;
 	std::vector<IObserver*> _observers;
-	int _number;
+	size_t _number;
 };
